requirements.txt
streamlit
pandas
numpy
matplotlib
fastapi
uvicorn
yfinance
streamlit_aggrid
python-dotenv 
requests